# Speed-up buildLayerGraph