Code complexity has risen up a bit with this change but layouting time will be reduced.

Here's the comparison of Worker Scripting time on Performance tab of Developer tools.

model | before (97c11fec36c49b699f9068f702286ecc17d46a04) [ms] | after time [ms]
------|------------:|----------------:
[EfficientDet-d0.onnx.zip](https://github.com/lutzroeder/netron/files/7786937/EfficientDet-d0.onnx.zip) | 6628 | 2717
retinanet_resnet50_fpn_v2_Opset18.onnx | 596 | 570